requirements.txt
google_auth_oauthlib==0.4.4
pytz==2020.1
requests==2.24.0
pyttsx3==2.90
wikipedia==1.4.0
pyowm==3.2.0
youtube_dl==2021.4.7
SpeechRecognition==3.8.1
urllib3==1.25.11
python_vlc==3.0.12117
beautifulsoup4==4.9.3
google_api_python_client==2.2.0
protobuf==3.15.8
